+++ spec/models/user_spec.rb
@@ -927,6 +927,54 @@ describe User, :type => :model do
     end
   end
 
+  describe "#send_welcome_message" do
+    let(:user) { FactoryGirl.create(:user) }
+    let(:podmin) { FactoryGirl.create(:user) }
+
+    context "with welcome message enabled" do
+      before do
+        AppConfig.settings.welcome_message.enabled = true
+      end
+
+      it "should send welcome message from podmin account" do
+        AppConfig.admins.account = podmin.username
+        expect {
+          user.send_welcome_message
+        }.to change(user.conversations, :count).by(1)
+        expect(user.conversations.first.author.owner.username).to eq podmin.username
+      end
+
+      it "should send welcome message text from config" do
+        AppConfig.admins.account = podmin.username
+        AppConfig.settings.welcome_message.text = "Hello %{username}, welcome!"
+        user.send_welcome_message
+        expect(user.conversations.first.messages.first.text).to eq "Hello #{user.username}, welcome!"
+      end
+
+      it "should use subject from config" do
+        AppConfig.settings.welcome_message.subject = "Welcome Message"
+        AppConfig.admins.account = podmin.username
+        user.send_welcome_message
+        expect(user.conversations.first.subject).to eq "Welcome Message"
+      end
+
+      it "should send no welcome message if no podmin is specified" do
+        AppConfig.admins.account = ""
+        user.send_welcome_message
+        expect(user.conversations.count).to eq 0
+      end
+    end
+
+    context "with welcome message disabled" do
+      it "shouldn't send a welcome message" do
+        AppConfig.settings.welcome_message.enabled = false
+        AppConfig.admins.account = podmin.username
+        user.send_welcome_message
+        expect(user.conversations.count).to eq 0
+      end
+    end
+  end
+
   context "close account" do
     before do
       @user = bob