+++ Changelog.md
@@ -12,6 +12,7 @@
 
 ## Features
 * Add configuration options for some debug logs [#6090](https://github.com/diaspora/diaspora/pull/6090)
+* Send new users a welcome message from the podmin [#6128](https://github.com/diaspora/diaspora/pull/6128)
 
 # 0.5.1.1